+++ packages/@expo/metro-config/src/customizeFrame.ts
@@ -108,6 +108,22 @@ export function getDefaultCustomizeFrame(): CustomizeFrameFunc {
         collapse = true;
       } else if (frame.file === '<native>') {
         collapse = true;
+      } else if (
+        // Some internal component stacks often don't have a file name.
+        frame.file === '<anonymous>' &&
+        frame.methodName &&
+        [
+          // React
+          'Suspense',
+          // React Native
+          'RCTView',
+          // React Native Screens
+          'RNSScreen',
+          'RNSScreenContentWrapper',
+          'RNSScreenNavigationContainer',
+        ].includes(frame.methodName)
+      ) {
+        collapse = true;
       }
     }