+++ packages/@expo/metro-config/CHANGELOG.md
@@ -12,6 +12,8 @@
 
 ### 💡 Others
 
+- Collapse more unused stacks.
+
 ## 0.20.4 — 2025-04-14
 
 _This version does not introduce any user-facing changes._